Python3 (#15)
* Fix the Activity to use Python 3 and to account for the changes in the
archive.org advanced search and in the kinds of files archive.org has.
(They no longer generate DJVUs and EPUBs for all their titles).

* Convert to Python 3.

* Add epub back as a download option.

* Remove references to old style toolbar.   These have not been needed for years.

* Fix bug where the table listing book info remains sensitive = False if the book path cannot be downloaded.

* Use https: for downloading books.

* Update to work with Python 3.

* Change search query to list books published in 1940 and earlier for not
every book that can be downloaded and they don't have a good way to limit results

* Improve the book search so that books that cannot be downloaded are
mostly omitted. Use the XML file list to get the actual name of the file
to be downloaded, instead of assuming that it is always the identifier
plus a file type suffix. In the book info pane, show what files are
available.

* Fix bug that prevented downloading djvus and epubs, update .pot.

---------

Co-authored-by: jamessimmons <jamessimmons@unknownF4390913EB01>
Co-authored-by: jdsimmon <jdsimmon@IT-LDR-SIMMONJ.walgreenco.net>